fix: adjust share button spacing, merge verse share link into share flow
